Updated move base example. Still failing when building
