+++ example_workspace/src/move_base_app/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Sun Feb 05 19:39:12 IST 2017
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-4.6-all.zip
