--- example_workspace/src/move_base_app/jni/Android.mk.in
@@ -1,15 +1,0 @@
-LOCAL_PATH := $(call my-dir)
-
-include $(CLEAR_VARS)
-
-LOCAL_MODULE    := move_base_app
-LOCAL_SRC_FILES := src/test_move_base.cpp
-LOCAL_C_INCLUDES := $(LOCAL_PATH)/include
-LOCAL_LDLIBS := -landroid
-LOCAL_STATIC_LIBRARIES := android_native_app_glue roscpp_android_ndk
-LOCAL_WHOLE_STATIC_LIBRARIES := liblayers libdwa_local_planner libclear_costmap_recovery librotate_recovery libglobal_planner libnavfn libtrajectory_planner_ros libcarrot_planner libmove_slow_and_clear
-
-include $(BUILD_SHARED_LIBRARY)
-
-$(call import-module,android/native_app_glue)
-$(call import-module,roscpp_android_ndk)
