--- example_workspace/src/move_base_app/jni/Application.mk
@@ -1,4 +1,0 @@
-#NDK_TOOLCHAIN_VERSION=4.4.3
-APP_STL := gnustl_static
-APP_PLATFORM := android-14
-APP_ABI := armeabi-v7a
