+++ example_workspace/src/move_base_app/jni/CMakeLists.txt
@@ -1,0 +1,50 @@
+#
+# Copyright (C) The Android Open Source Project
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#      http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+cmake_minimum_required(VERSION 3.4.1)
+project(move_base_app)
+
+find_package(catkin REQUIRED COMPONENTS roscpp move_base)
+
+include_directories(${catkin_INCLUDE_DIRS})
+
+# build native_app_glue as a static lib
+set(${CMAKE_C_FLAGS}, "${CMAKE_C_FLAGS}")
+add_library(native_app_glue STATIC
+    ${ANDROID_NDK}/sources/android/native_app_glue/android_native_app_glue.c)
+
+# now build app's shared lib
+set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=gnu++11 -Wall -Werror")
+
+# Export ANativeActivity_onCreate(),
+# Refer to: https://github.com/android-ndk/ndk/issues/381.
+set(CMAKE_SHARED_LINKER_FLAGS
+    "${CMAKE_SHARED_LINKER_FLAGS} -u ANativeActivity_onCreate")
+
+add_library(${PROJECT_NAME} SHARED 
+    src/main/cpp/test_move_base.cpp
+)
+
+target_include_directories(${PROJECT_NAME} PRIVATE
+    ${ANDROID_NDK}/sources/android/native_app_glue)
+
+# add lib dependencies
+target_link_libraries(${PROJECT_NAME}
+    android
+    native_app_glue
+    ${catkin_LIBRARIES}
+    log
+)
