+++ example_workspace/src/move_base_app/jni/build.gradle
@@ -1,0 +1,43 @@
+apply plugin: 'com.android.application'
+
+android {
+    compileSdkVersion 28
+
+    defaultConfig {
+        applicationId = 'com.example.move_base_app'
+        minSdkVersion 24
+        targetSdkVersion 24
+        externalNativeBuild {
+            cmake {
+                arguments   '-DANDROID_STL=c++_shared',
+                            '-DCMAKE_FIND_ROOT_PATH=' + System.getenv('TARGET_DIR'),
+                            '-DCMAKE_PREFIX_PATH=' + System.getenv('TARGET_DIR'),
+                            '-DBUILD_TESTING=OFF',
+                            '-DCATKIN_ENABLE_TESTING=OFF'
+            }
+        }
+        ndk {
+            // Specifies the ABI configurations of your native
+            // libraries Gradle should build and package with your APK.
+            abiFilters 'arm64-v8a'
+        }
+    }
+    buildTypes {
+        release {
+            minifyEnabled false
+            proguardFiles getDefaultProguardFile('proguard-android.txt'),
+                    'proguard-rules.pro'
+        }
+    }
+    externalNativeBuild {
+        cmake {
+            path 'CMakeLists.txt'
+        }
+    }
+}
+
+dependencies {
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    implementation 'com.android.support:appcompat-v7:28.0.0'
+    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
+}
