+++ example_workspace/src/move_base_app/jni/src/main/res/values/strings.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="app_name">MoveBase</string>
+</resources>
