+++ example_workspace/src/move_base_app/package.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0"?>
+<package format="2">
+  <name>move_base_app</name>
+  <version>0.0.1</version>
+  <description>Hello word example with ros on android</description>
+
+  <maintainer email="ivanpauno@ekumenlabs.com">Ivan Paunovic</maintainer>
+
+  <!-- One license tag required, multiple allowed, one license per tag -->
+  <!-- Commonly used license strings: -->
+  <!--   BSD, MIT, Boost Software License, GPLv2, GPLv3, LGPLv2.1, LGPLv3 -->
+  <license>TODO</license>
+
+  <buildtool_depend>catkin</buildtool_depend>
+  <depend>roscpp</depend>
+  <depend>move_base</depend>
+
+</package>
