+++ example_workspace/src/move_base_app/settings.gradle
@@ -1,0 +1,2 @@
+include ':jni'
+
